Updating 23.7 with latest features from WNPRC (#445)
* 22.11 fb water lixit improvements (#437)

* Adding Lixit animals to the waterCalendar view.

* Modifying water alerts to include alert for supervisors. Moving functions to base notification

* Adding location using labkey study lookup in .qview.xml

* Adding Lixit to TotalWater to the row in the event calendar

* Using conditionattime to modify the water Total details panel.

* Adding location to the notification.

* Adding last day for condition and improving display for taskid

* Adding demographics column to show water condition trough-tout the system.

* Improving titles on the mostRecentWater demographics integration.

* Adding additional color to water calendar and displaying remarks to the Total Water by Date report.

* Fixing problem with tomcat name and improve build.gradle (#439)

* Making a generic docker build

* Adding some more echo in the build script

* Adding LabKey version for SNAPSHOT

* Remove version num, adding support for develop bch

* Improving support for develop branch

* adding post_push to send to different repos

* improving post_push to send image to other repos

* adding support for different Tomcat versions

* Moving echo closer to delete command

* Matching Git_Hub_Tag with docker_tag

* Adding logic to docker build script

* Switching to use Curl

* printing teamcity URL

* Moving variables outside if statement

* Fixing problem with LK_Version

* Adding echo to see value of LKVer

* Closing quotes for empty string

* Changing number of build hook

* updating build version

* moving BASH_REMATCH inside if statement

* Adding more echos

* increasing build version

* Moving URL logic to build hook

* Removing echo from build script

* Updating gradle build to work with docker hooks

* updating gradlew and adding version to lk build

* Adding gradle tasks for docker images and updating build command

* adding gradle properties to the docker folder

* updating gradle wrapper

* Opening docker always

* Openning docker in a linux command

* Adding setenv.sh to setup memory for Tomcat process

* Updating compose file to work with compose v2

* Removing all docker-compose files

* Changing variables name for perl backup script

* adding setenv.sh file

* Removing space between file and location

* removing distination file from copy

* Chaning name for default.env variables

* Changing varaible names to adhere to new .env names

* Updating build.gradle to download from Docker Hub

* Removing Echo and updating version

* adding no cache to build LabKey

* Adding underscore to tomcat image

* Adding full path to COPY command

* Adding setenv.sh to white list

* Changing compose file to use secrets to improve docker security

* Moving default secrets

* changing file in secrets

* removing volumes from mailserver

* adding dependsOn to gradle task

* updating JDK and OS

* Adding cert to postfix via secrets

* Updating default.env

* Changing to start service in Ubuntu

* adding gradle support for check docker in linux

* Modifying comandline for Linux

* Changing login command

* Modify all the gradle tasks to the new lazy loading method.

* Adding blank properties for DockerHub variables

* Improving login to dockerhub and bash to login

* removing password to use token

* Removing reference to password

* Migrating new compose  to the develop folder

* Updated wnprcReport.js with new script location on line 1167.

* Reverted code in wnprcReports.js back to original file location.

* Adding password to be used with plugin.

* maintaing postgres.conf from production

* Updating postgresql client for backup

* Update to postgresql.conf to version 15

* Updating working memory

* Updating documentation to use gradle tasks.

* moving postfix variables together.

* Adding information about the new variable in .env

* Adding download command for tomcat

* updating postgres.conf file

* adding remotes to r script

* Moving underscore to build hook to work snapshot

* fixing names of base tomcat image

---------


* Adding filter to show null for Today at WNPRC (#441)

* removing last day from previous water order when entering a lixit

* Adding filter to show null in the today at wnprc report

* Fix to the water calendar display.

* Fixing problem with full history when displaying remarks from Water given. (#444)

---------
