+++ WNPRC_EHR/resources/queries/study/WaterScheduleWithWeight/.qview.xml
@@ -1,0 +1,32 @@
+<customView xmlns="http://labkey.org/data/xml/queryCustomView">
+    <columns>
+        <column name="lsid" />
+        <column name="objectIdCoalesced"/>
+        <column name="waterOrderObjectId"/>
+        <column name="projectCoalesced"/>
+        <column name="Id" />
+        <column name="date" />
+        <column name="dateOrdered"/>
+        <column name="location"/>
+        <column name="volume" />
+        <column name="provideFruit"/>
+        <column name="provideFruitTitle"/>
+        <column name="dataSource"/>
+        <column name="assignedToCoalesced"/>
+        <column name="assignedToTitleCoalesced"/>
+        <column name="frequencyCoalesced"/>
+        <column name="frequencyMeaningCoalesced"/>
+        <column name="displaytimeofday"/>
+        <column name="rawDate"/>
+        <column name="mlsPerKg"/>
+        <column name="conditionAtTime"/>
+        <column name="qcstate" />
+        <column name="taskid" />
+    </columns>
+    <sorts>
+        <sort column="date" descending="false"/>
+        <sort column="location" />
+        <sort column="Id"/>
+
+    </sorts>
+</customView>
