+++ WNPRC_EHR/resources/queries/study/waterScheduleCoalesced/Treatments.qview.xml
@@ -10,7 +10,7 @@
         <column name="waterSource"/>
         <column name="assignedTo"/>
         <column name="provideFruit"/>
-        <column name="qcstate/Label">
+        <column name="qcstate">
             <properties>
                 <property name="columnTitle" value="Status"/>
             </properties>
@@ -26,6 +26,6 @@
     <filters>
         <filter column="dateOrdered" operator="dategte" value="-30d"/>
         <filter column="Id/dataset/demographics/calculated_status" operator="eq" value="Alive"/>
-        <filter column="qctate/Label" operator="neqornull" value="Completed"/>
+        <filter column="qcstate" operator="neqornull" value="Completed"/>
     </filters>
 </customView>
