+++ WNPRC_EHR/resources/queries/study/waterScheduledAnimalWithOutEntries.sql
@@ -1,7 +1,18 @@
 PARAMETERS(CheckDate TIMESTAMP)
 
-SELECT DISTINCT(id), CheckDate AS date, project AS project FROM
-    (SELECT a.id, a.MostRecentWaterCondition, a.MostRecentWaterConditionDate, a.project FROM study.demographicsMostRecentWaterCondition a
+SELECT
+    DISTINCT(id),
+    CheckDate AS date,
+    project AS project,
+    location as location
+FROM
+    (SELECT
+        a.id,
+        a.MostRecentWaterCondition,
+        a.MostRecentWaterConditionDate,
+        a.project,
+        a.location
+    FROM study.demographicsMostRecentWaterCondition a
      WHERE NOT EXISTS (
          SELECT 1 FROM study.waterTotalByDate b
                        WHERE a.id = b.id AND b.date = CheckDate AND b.TotalWater IS NOT NULL
