+++ WNPRC_EHR/resources/queries/study/waterScheduledAnimalWithOutEntries/.qview.xml
@@ -1,0 +1,11 @@
+<customView xmlns="http://labkey.org/data/xml/queryCustomView">
+    <columns>
+        <column name="Id" />
+        <column name="location"/>
+        <column name="date" />
+        <column name="project"/>
+    </columns>
+    <sorts>
+        <sort column="date" descending="true"/>
+    </sorts>
+</customView>
