+++ WNPRC_EHR/resources/queries/study/waterTotalByDate.sql
@@ -9,46 +9,64 @@ SELECT
        cal.targetdate AS date,
        waterSummary.TotalWater,
        waterSummary.provideFruit,
-       waterSummary.remark,
+       waterSummary.remarksConcat,
        waterSummary.volumeGivenInLabSub,
        waterSummary.volumeGivenInCage,
        waterSummary.volumeGivenInImage,
        waterSummary.volumeGivenInProcedure,
        waterSummary.performedConcat,
        odrwc.currentWaterCondition,
+       odrwc.latestConditionDate,
        waterSummary.project,
-       waterSummary.qcstate,
-       waterSummary.curRoom,
-       waterSummary.curCage
+       waterSummary.qcstate
 
 FROM ehr_lookups.calendar cal
- LEFT OUTER JOIN (SELECT drwc.Id AS Id,
-                 MIN(drwc.date) AS firstDate,
-                 (SELECT wsainner.condition
-                  FROM study.waterScheduledAnimals wsainner
-                  WHERE drwc.id = wsainner.id
-                  ORDER BY wsainner.date DESC LIMIT 1) AS currentWaterCondition
-
+ LEFT OUTER JOIN (
+    SELECT
+        drwc.Id AS Id,
+        MIN(drwc.date) AS firstDate,
+        (SELECT
+            wsainner.condition
+            FROM study.waterScheduledAnimals wsainner
+            WHERE drwc.id = wsainner.id
+            ORDER BY wsainner.date DESC LIMIT 1) AS currentWaterCondition,
+        (SELECT
+            wsa2inner.date
+            FROM study.waterScheduledAnimals wsa2inner
+            WHERE drwc.id = wsa2inner.id
+            ORDER BY wsa2inner.date DESC LIMIT 1) AS latestConditionDate
         FROM study.waterScheduledAnimals drwc
-        GROUP BY Id) odrwc
-        ON cal.targetdate >= CAST (odrwc.firstDate AS DATE) AND cal.targetdate <= curdate()
-LEFT OUTER JOIN(
-    SELECT Id AS innerId,
-    CAST(iwg.date as DATE) as date,
-    COALESCE (SUM(CASE WHEN iwg.location = 'lab' THEN iwg.volume ELSE 0 END),0) AS volumeGivenInLabSub,
-    COALESCE (SUM(CASE WHEN iwg.location = 'animalRoom' THEN iwg.volume ELSE 0 END),0) AS volumeGivenInCage,
-    COALESCE (SUM(CASE WHEN iwg.location = 'imaging' THEN iwg.volume ELSE 0 END), 0) AS volumeGivenInImage,
-    COALESCE (SUM(CASE WHEN iwg.location = 'procedureRoom' THEN iwg.volume ELSE 0 END), 0) AS volumeGivenInProcedure,
-    COALESCE(SUM(iwg.volume), 0) AS TotalWater,
-    COALESCE(GROUP_CONCAT(iwg.provideFruit.title,'; '),'') AS provideFruit,
-    COALESCE(GROUP_CONCAT(iwg.remarks,'; '),'') AS remark,
-    COALESCE(GROUP_CONCAT(iwg.performedby, ';'),'') AS performedConcat,
-    MAX(iwg.project) AS project,
-    COALESCE(MAX(iwg.qcstate),1) AS qcstate,
-    MAX(iwg.id.curLocation.room) as curRoom,
-    MAX(iwg.id.curLocation.cage) as curCage,
-    FROM study.waterGiven iwg WHERE qcstate.label = 'Completed'
-    GROUP BY iwg.Id, CAST(iwg.date AS DATE)
-    )waterSummary
-ON CAST(cal.targetdate AS DATE) = waterSummary.date AND waterSummary.innerId = Id
-WHERE cal.targetdate <= curdate() AND Id IS NOT NULL
+        GROUP BY Id
+ ) odrwc
+ ON cal.targetdate >= CAST (odrwc.firstDate AS DATE)
+ AND cal.targetdate <= curdate()
+ AND ( Id.death.date IS NULL OR cal.targetdate <= CAST(Id.death.date AS DATE) )
+
+ LEFT OUTER JOIN (
+    SELECT
+        Id AS innerId,
+        CAST(iwg.date as DATE) as date,
+        COALESCE (SUM(CASE WHEN iwg.location = 'lab' THEN iwg.volume ELSE 0 END),0) AS volumeGivenInLabSub,
+        COALESCE (SUM(CASE WHEN iwg.location = 'animalRoom' THEN iwg.volume ELSE 0 END),0) AS volumeGivenInCage,
+        COALESCE (SUM(CASE WHEN iwg.location = 'imaging' THEN iwg.volume ELSE 0 END), 0) AS volumeGivenInImage,
+        COALESCE (SUM(CASE WHEN iwg.location = 'procedureRoom' THEN iwg.volume ELSE 0 END), 0) AS volumeGivenInProcedure,
+        COALESCE(SUM(iwg.volume), 0) AS TotalWater,
+        COALESCE(GROUP_CONCAT(iwg.provideFruit.title,'; '),'') AS provideFruit,
+        COALESCE(GROUP_CONCAT(iwg.remarks,'; '),'') AS remarksConcat,
+        COALESCE(GROUP_CONCAT(iwg.performedby, ';'),'') AS performedConcat,
+        MAX(iwg.project) AS project,
+        COALESCE(MAX(iwg.qcstate),1) AS qcstate,
+    FROM study.waterGiven iwg
+    WHERE
+        qcstate.label = 'Completed'
+        AND ( iwg.Id.death.date IS NULL OR CAST(iwg.date as DATE) <= iwg.Id.death.date )
+    GROUP BY
+        iwg.Id,
+        CAST(iwg.date AS DATE)
+ )waterSummary
+ON
+    CAST(cal.targetdate AS DATE) = waterSummary.date
+    AND waterSummary.innerId = Id
+WHERE
+    cal.targetdate <= curdate()
+    AND Id IS NOT NULL
