+++ WNPRC_EHR/resources/queries/study/waterTotalByDate/.qview.xml
@@ -1,6 +1,7 @@
 <customView xmlns="http://labkey.org/data/xml/queryCustomView">
     <columns>
         <column name="Id" />
+        <column name="Id/curLocation/location"/>
         <column name="date" />
         <column name="TotalWater" />
         <column name="provideFruit"/>
@@ -12,6 +13,7 @@
         <column name="volumeGivenInProcedure" />
         <column name="performedConcat" />
         <column name="currentWaterCondition" />
+        <column name="latestConditionDate" />
         <column name="project"/>
         <column name="qcstate" />
     </columns>
