+++ WNPRC_EHR/src/org/labkey/wnprc_ehr/table/WNPRC_EHRCustomizer.java
@@ -388,6 +388,11 @@ public class WNPRC_EHRCustomizer extends AbstractTableCustomizer
             col20.setLabel("PrimateId");
             col20.setDescription("Unique PrimateID column to be shared across all datasets");
             ds.addColumn(col20);
+
+            BaseColumnInfo col21 = getWrappedIdCol(us, ds, "mostRecentWaterCondition", "demographicsMostRecentWaterCondition" );
+            col21.setLabel("Most Recent Water Condition");
+            col21.setDescription("Most Recent Water condition for animals in Water restricted protocols");
+            ds.addColumn(col21);
         }
 
         if (ds.getColumn("totalOffspring") == null)
