+++ docker/development/default.env
@@ -10,7 +10,7 @@ LK_FILES_DIR=./data/files
 LK_JAVA_OPTS=
 LK_LOG_DIR=./logs/labkey
 LK_MASTER_ENCRYPTION_KEY=fakePassword
-LK_MODULE_DEPLOY_DIR=../build/modules
+LK_MODULE_DEPLOY_DIR=./modules
 LK_MSQL_PASS=
 LK_MSQL_URL=
 LK_MSQL_USER=labkey
@@ -26,3 +26,7 @@ NGINX_SERVER_CER_FILE=./nginx/cert.pem
 NGINX_SERVER_KEY_FILE=./nginx/key.pem
 NGINX_SERVER_PASS_FILE=./nginx/key.pass
 NGINX_WELL_KNOWN_DIR=./data/well-known
+PG_NAME=labkey_new_db
+PG_PASS=
+PG_PORT=9081
+PG_USER=postgres
