+++ docker/gradle.properties
@@ -90,3 +90,7 @@ gsonVersion=2.8.9
 dockerhubUsername=
 dockerhubPassword=
 dockertokenpath=
+
+labkeyYear=22
+labkeyMonth=11
+labkeyFB=
